feat: Windows SSH agent support
Configures the OpenSSH Agent on the Windows host to be enabled and prompts the user to provide the password for a `homelab` SSH key if it is present. This also installs socat and npiperelay which allows for sharing the OpenSSH named pipe from Windows into the WSL instance. This enables the user to provide the credentials for their SSH key once and they are stored securely in the Windows credential manager and they may be used to SSH from within WSL.